+++ 4k.c
@@ -200,19 +200,7 @@ typedef struct [[nodiscard]] {
   return ts.tv_sec * 1000 + ts.tv_nsec / 1000000;
 }
 
-#ifdef ASSERTS
-#define assert(condition)                                                      \
-  if (!(condition)) {                                                          \
-    printf("Assert failed on line %i: ", __LINE__);                            \
-    puts(#condition);                                                          \
-    _sys(H(2, 6, 1), H(2, 6, 0), H(2, 6, 0), H(2, 6, 60));                     \
-  }
 #else
-#define assert(condition)
-#endif
-
-#else
-#include <assert.h>
 #include <stdbool.h>
 #include <stdio.h>
 #include <stdlib.h>
@@ -266,18 +254,29 @@ enum [[nodiscard]] { None, Pawn, Knight, Bishop, Rook, Queen, King };
 
 typedef struct [[nodiscard]] {
   G(5, u8 promo;)
-  G(5, u8 takes_piece;)
   G(5, u8 from; u8 to;)
+  G(5, u8 takes_piece;)
 } Move;
 
 typedef struct [[nodiscard]] {
   G(6, u64 ep;)
   G(6, u64 pieces[7];)
   G(6, u64 colour[2];)
-  G(7, bool castling[4];)
   G(7, bool flipped;)
+  G(7, bool castling[4];)
   G(7, u8 padding[11];)
 } Position;
+
+#ifdef ASSERTS
+#define assert(condition)                                                      \
+  if (!(condition)) {                                                          \
+    printf("Assert failed on line %i: ", __LINE__);                            \
+    puts(#condition);                                                          \
+    exit_now();                                                                \
+  }
+#else
+#define assert(condition)
+#endif
 
 [[nodiscard]] S(1) bool move_string_equal(G(8, const char *restrict lhs),
                                           G(8, const char *restrict rhs)) {
@@ -393,7 +392,7 @@ G(
 
 G(
     22,
-    S(1) void swapu32(G(23, u32 *const lhs), G(23, u32 *const rhs)) {
+    S(1) void swapu32(G(23, u32 *const rhs), G(23, u32 *const lhs)) {
       const u32 temp = *lhs;
       *lhs = *rhs;
       *rhs = temp;
@@ -480,18 +479,18 @@ G(
 
 G(
     32, S(1) void flip_pos(Position *const restrict pos) {
-      G(
-          37, for (i32 i = 0; i < 2; i++) {
-            swapbool(G(38, &pos->castling[i + 2]), G(38, &pos->castling[i]));
-          })
+      G(37, swapu64(G(38, &pos->colour[0]), G(38, &pos->colour[1]));)
 
-      G(37, pos->flipped ^= 1;)
-      G(37, swapu64(G(39, &pos->colour[0]), G(39, &pos->colour[1]));)
       G(
           37, // Hack to flip the first 10 bitboards in Position.
-              // Technically UB but works in GCC 14.2
+          // Technically UB but works in GCC 14.2
           u64 *pos_ptr = (u64 *)pos;
           for (i32 i = 0; i < 10; i++) { pos_ptr[i] = flip_bb(pos_ptr[i]); })
+      G(37, pos->flipped ^= 1;)
+      G(
+          37, for (i32 i = 0; i < 2; i++) {
+            swapbool(G(39, &pos->castling[i + 2]), G(39, &pos->castling[i]));
+          })
     })
 
 G(
@@ -515,7 +514,7 @@ G(
       return moves;
     })
 
-S(1) i32 find_in_check(const Position *restrict pos) {
+S(0) i32 find_in_check(const Position *restrict pos) {
   return is_attacked(H(33, 2, pos),
                      H(33, 2, pos->colour[0] & pos->pieces[King]));
 }
@@ -689,44 +688,53 @@ enum { max_moves = 218 };
   G(64, const u64 all = pos->colour[0] | pos->colour[1];)
   G(64, const Move *start = movelist;)
   G(64, const u64 to_mask = only_captures ? pos->colour[1] : ~pos->colour[0];)
-  if (!only_captures) {
+  G(
+      65, // PAWN PROMOTIONS
+      if (!only_captures) {
+        movelist = generate_pawn_moves(
+            H(62, 2, pos), H(62, 2, movelist),
+            H(62, 2,
+              north(north(pos->colour[0] & pos->pieces[Pawn] & 0xFF00) & ~all) &
+                  ~all),
+            H(62, 2, -16));
+      })
+  G(65, // PAWN DOUBLE MOVES
     movelist = generate_pawn_moves(
-        H(62, 2, pos), H(62, 2, movelist),
-        H(62, 2,
-          north(north(pos->colour[0] & pos->pieces[Pawn] & 0xFF00) & ~all) &
-              ~all),
-        H(62, 2, -16));
-  }
-  movelist = generate_pawn_moves(
-      H(62, 3, pos), H(62, 3, movelist),
-      H(62, 3,
-        north(pos->colour[0] & pos->pieces[Pawn]) & ~all &
-            (only_captures ? 0xFF00000000000000ull : ~0ull)),
-      H(62, 3, -8));
-  movelist = generate_pawn_moves(
-      H(62, 4, pos), H(62, 4, movelist),
-      H(62, 4,
-        nw(pos->colour[0] & pos->pieces[Pawn]) & (pos->colour[1] | pos->ep)),
-      H(62, 4, -7));
-  movelist = generate_pawn_moves(
-      H(62, 5, pos), H(62, 5, movelist),
-      H(62, 5,
-        ne(pos->colour[0] & pos->pieces[Pawn]) & (pos->colour[1] | pos->ep)),
-      H(62, 5, -9));
-  if (G(65, !only_captures) && G(65, pos->castling[0]) &&
-      G(65, !(all & 0x60ull)) &&
-      G(66, !is_attacked(H(33, 3, pos), H(33, 3, 1ULL << 5))) &&
-      G(66, !is_attacked(H(33, 4, pos), H(33, 4, 1ULL << 4)))) {
-    *movelist++ =
-        (Move){.from = 4, .to = 6, .promo = None, .takes_piece = None};
-  }
-  if (G(67, !only_captures) && G(67, pos->castling[1]) &&
-      G(67, !(all & 0xEull)) &&
-      G(68, !is_attacked(H(33, 5, pos), H(33, 5, 1ULL << 3))) &&
-      G(68, !is_attacked(H(33, 6, pos), H(33, 6, 1ULL << 4)))) {
-    *movelist++ =
-        (Move){.from = 4, .to = 2, .promo = None, .takes_piece = None};
-  }
+        H(62, 3, pos), H(62, 3, movelist),
+        H(62, 3,
+          north(pos->colour[0] & pos->pieces[Pawn]) & ~all &
+              (only_captures ? 0xFF00000000000000ull : ~0ull)),
+        H(62, 3, -8));)
+  G(
+      65, // LONG CASTLE
+      if (G(66, !only_captures) && G(66, pos->castling[0]) &&
+          G(66, !(all & 0x60ull)) &&
+          G(67, !is_attacked(H(33, 3, pos), H(33, 3, 1ULL << 5))) &&
+          G(67, !is_attacked(H(33, 4, pos), H(33, 4, 1ULL << 4)))) {
+        *movelist++ =
+            (Move){.from = 4, .to = 6, .promo = None, .takes_piece = None};
+      })
+  G(65, // PAWN EAST CAPTURES
+    movelist = generate_pawn_moves(
+        H(62, 4, pos), H(62, 4, movelist),
+        H(62, 4,
+          ne(pos->colour[0] & pos->pieces[Pawn]) & (pos->colour[1] | pos->ep)),
+        H(62, 4, -9));)
+  G(
+      65, // SHORT CASTLE
+      if (G(68, !only_captures) && G(68, pos->castling[1]) &&
+          G(68, !(all & 0xEull)) &&
+          G(69, !is_attacked(H(33, 5, pos), H(33, 5, 1ULL << 3))) &&
+          G(69, !is_attacked(H(33, 6, pos), H(33, 6, 1ULL << 4)))) {
+        *movelist++ =
+            (Move){.from = 4, .to = 2, .promo = None, .takes_piece = None};
+      })
+  G(65, // PAWN WEST CAPTURES
+    movelist = generate_pawn_moves(
+        H(62, 5, pos), H(62, 5, movelist),
+        H(62, 5,
+          nw(pos->colour[0] & pos->pieces[Pawn]) & (pos->colour[1] | pos->ep)),
+        H(62, 5, -7));)
   movelist = generate_piece_moves(H(58, 2, to_mask), H(58, 2, movelist),
                                   H(58, 2, pos));
 
@@ -856,34 +864,34 @@ static void get_fen(Position *restrict pos, char *restrict fen) {
 
 typedef struct [[nodiscard]] __attribute__((packed)) {
   i16 material[6];
-  H(69, 1,
-    H(70, 1, i8 king_attacks[5];) H(70, 1, i8 mobilities[5];)
-        H(70, 1, i8 passed_pawns[6];) H(70, 1, i8 passed_blocked_pawns[6];)
-            H(70, 1, i8 tempo;))
-  H(69, 1,
-    H(71, 1, i8 bishop_pair;) H(71, 1, u8 pawn_attacked_penalty[2];)
-        H(71, 1, i8 open_files[6];) H(71, 1, i8 pst_file[64];)
-            H(71, 1, i8 pst_rank[64];))
+  H(70, 1,
+    H(71, 1, i8 king_attacks[5];) H(71, 1, i8 passed_pawns[6];)
+        H(71, 1, i8 tempo;) H(71, 1, i8 passed_blocked_pawns[6];)
+            H(71, 1, i8 mobilities[5];))
+  H(70, 1,
+    H(72, 1, i8 bishop_pair;) H(72, 1, i8 open_files[6];)
+        H(72, 1, u8 pawn_attacked_penalty[2];) H(72, 1, i8 pst_rank[64];)
+            H(72, 1, i8 pst_file[64];))
 } EvalParams;
 
 typedef struct [[nodiscard]] __attribute__((packed)) {
   i32 material[6];
-  H(69, 2,
-    H(70, 2, i32 king_attacks[5];) H(70, 2, i32 mobilities[5];)
-        H(70, 2, i32 passed_pawns[6];) H(70, 2, i32 passed_blocked_pawns[6];)
-            H(70, 2, i32 tempo;))
-  H(69, 2,
-    H(71, 2, i32 bishop_pair;) H(71, 2, i32 pawn_attacked_penalty[2];)
-        H(71, 2, i32 open_files[6];) H(71, 2, i32 pst_file[64];)
-            H(71, 2, i32 pst_rank[64];))
+  H(70, 2,
+    H(71, 2, i32 king_attacks[5];) H(71, 2, i32 passed_pawns[6];)
+        H(71, 2, i32 tempo;) H(71, 2, i32 passed_blocked_pawns[6];)
+            H(71, 2, i32 mobilities[5];))
+  H(70, 2,
+    H(72, 2, i32 bishop_pair;) H(72, 2, i32 open_files[6];)
+        H(72, 2, i32 pawn_attacked_penalty[2];) H(72, 2, i32 pst_rank[64];)
+            H(72, 2, i32 pst_file[64];))
 
 } EvalParamsMerged;
 
-G(72,
+G(73,
   __attribute__((aligned(8))) S(1) const i8 phases[] = {0, 0, 1, 1, 2, 4, 0};)
-G(72, S(0) EvalParamsMerged eval_params;)
+G(73, S(0) EvalParamsMerged eval_params;)
 
-G(72, S(1) const EvalParams mg = ((EvalParams){
+G(73, S(1) const EvalParams mg = ((EvalParams){
           .material = {67, 266, 270, 357, 768, 0},
           .pst_rank =
               {
@@ -912,16 +920,16 @@ G(72, S(1) const EvalParams mg = ((EvalParams){
           .pawn_attacked_penalty = {-16, -128},
           .tempo = 17});)
 
-G(72, __attribute__((aligned(8))) S(1)
+G(73, __attribute__((aligned(8))) S(1)
           const i16 max_material[] = {0, 86, 398, 392, 706, 1340};)
 
 G(
-    72, [[nodiscard]] S(1) i32 combine_eval_param(H(73, 1, const i32 mg_val),
-                                                  H(73, 1, const i32 eg_val)) {
-      return G(74, mg_val) + G(74, (eg_val << 16));
+    73, [[nodiscard]] S(1) i32 combine_eval_param(H(74, 1, const i32 mg_val),
+                                                  H(74, 1, const i32 eg_val)) {
+      return G(75, mg_val) + G(75, (eg_val << 16));
     })
 
-G(72, S(1) const EvalParams eg = ((EvalParams){
+G(73, S(1) const EvalParams eg = ((EvalParams){
           .material = {86, 398, 392, 706, 1340, 0},
           .pst_rank =
               {
@@ -950,112 +958,83 @@ G(72, S(1) const EvalParams eg = ((EvalParams){
           .pawn_attacked_penalty = {-10, -128},
           .tempo = 7});)
 
-S(1) void init() {
-  // INIT DIAGONAL MASKS
-  G(
-      46, for (i32 sq = 0; sq < 64; sq++) {
-        const u64 bb = 1ULL << sq;
-        diag_mask[sq] = G(75, ray(H(15, 4, 0), H(15, 4, ~0x8080808080808080ull),
-                                  H(15, 4, bb), H(15, 4, -9))) | // Northeast
-                        G(75, ray(H(15, 5, 0), H(15, 5, ~0x101010101010101ull),
-                                  H(15, 5, bb), H(15, 5, 9))); // Southwest
-      })
-
-  G(
-      46, // MERGE MATERIAL VALUES
-      for (i32 i = 0; i < sizeof(mg.material) / sizeof(i16); i++) {
-        eval_params.material[i] = combine_eval_param(H(73, 2, mg.material[i]),
-                                                     H(73, 2, eg.material[i]));
-      })
-  G(
-      46, // MERGE NON-MATERIAL VALUES
-      for (i32 i = 0; i < sizeof(mg) - sizeof(mg.material); i++) {
-        // Technically writes past end of array
-        // But since the structs are packed, it works
-        const i32 offset = sizeof(mg.material);
-        ((i32 *)&eval_params)[offset / sizeof(*mg.material) + i] =
-            combine_eval_param(H(73, 3, ((i8 *)&mg)[offset + i]),
-                               H(73, 3, ((i8 *)&eg)[offset + i]));
-      })
-}
-
 S(1) i32 eval(Position *const restrict pos) {
-  G(77, i32 score = eval_params.tempo;)
-  G(77, i32 phase = 0;)
+  G(76, i32 score = eval_params.tempo;)
+  G(76, i32 phase = 0;)
 
   for (i32 c = 0; c < 2; c++) {
 
     G(
-        78, // BISHOP PAIR
-        if (count(G(79, pos->pieces[Bishop]) & G(79, pos->colour[0])) > 1) {
+        77, // BISHOP PAIR
+        if (count(G(78, pos->pieces[Bishop]) & G(78, pos->colour[0])) > 1) {
           score += eval_params.bishop_pair;
         })
 
-    G(78,
-      const u64 opp_pawns = G(80, pos->pieces[Pawn]) & G(80, pos->colour[1]);
-      const u64 attacked_by_pawns = G(81, se(opp_pawns)) | G(81, sw(opp_pawns));
-      const u64 no_passers = G(82, opp_pawns) | G(82, attacked_by_pawns);)
-    G(78, const u64 opp_king_zone = king(pos->colour[1] & pos->pieces[King]);)
-    G(78,
-      const u64 own_pawns = G(83, pos->pieces[Pawn]) & G(83, pos->colour[0]);)
+    G(77,
+      const u64 opp_pawns = G(79, pos->pieces[Pawn]) & G(79, pos->colour[1]);
+      const u64 attacked_by_pawns = G(80, se(opp_pawns)) | G(80, sw(opp_pawns));
+      const u64 no_passers = G(81, opp_pawns) | G(81, attacked_by_pawns);)
+    G(77,
+      const u64 own_pawns = G(82, pos->pieces[Pawn]) & G(82, pos->colour[0]);)
+    G(77, const u64 opp_king_zone = king(pos->colour[1] & pos->pieces[King]);)
 
     for (i32 p = Pawn; p <= King; p++) {
-      u64 copy = G(84, pos->colour[0]) & G(84, pos->pieces[p]);
+      u64 copy = G(83, pos->colour[0]) & G(83, pos->pieces[p]);
       while (copy) {
         const i32 sq = lsb(copy);
-        G(85, const int file = sq & 7;)
-        G(85, const int rank = sq >> 3;)
-        G(85, phase += phases[p];)
-        G(85, copy &= copy - 1;)
+        G(84, const int file = sq & 7;)
+        G(84, const int rank = sq >> 3;)
+        G(84, phase += phases[p];)
+        G(84, copy &= copy - 1;)
 
         G(
             62, // OPEN FILES / DOUBLED PAWNS
-            if ((G(86, north(0x101010101010101ULL << sq)) & G(86, own_pawns)) ==
+            if ((G(85, north(0x101010101010101ULL << sq)) & G(85, own_pawns)) ==
                 0) { score += eval_params.open_files[p - 1]; })
-
-        G(62, // SPLIT PIECE-SQUARE TABLES FOR FILE
-          score += eval_params.pst_file[(p - 1) * 8 + file];)
-        G(62, // SPLIT PIECE-SQUARE TABLES FOR RANK
-          score += eval_params.pst_rank[(p - 1) * 8 + rank];)
 
         G(62, // MATERIAL
           score += eval_params.material[p - 1];)
+        G(62, // SPLIT PIECE-SQUARE TABLES FOR FILE
+          score += eval_params.pst_file[(p - 1) * 8 + file];)
+
+        G(62, // SPLIT PIECE-SQUARE TABLES FOR RANK
+          score += eval_params.pst_rank[(p - 1) * 8 + rank];)
 
         G(
             62, if (p > Pawn) {
               G(
-                  87, // PIECES ATTACKED BY PAWNS
+                  86, // PIECES ATTACKED BY PAWNS
                   if (1ULL << sq & no_passers) {
                     score += eval_params.pawn_attacked_penalty[c];
                   })
 
-              G(87, const u64 mobility =
-                        G(88, get_mobility(H(40, 3, pos), H(40, 3, sq),
+              G(86, const u64 mobility =
+                        G(87, get_mobility(H(40, 3, pos), H(40, 3, sq),
                                            H(40, 3, p))) &
-                        G(88, G(89, ~attacked_by_pawns));
+                        G(87, G(88, ~pos->colour[0]));
 
-                G(90, // MOBILITY
+                G(89, // MOBILITY
                   score +=
-                  G(91, count(G(89, mobility) & G(89, ~pos->colour[0]))) *
-                  G(91, eval_params.mobilities[p - 2]);)
+                  G(90, count(G(88, mobility) & G(88, ~attacked_by_pawns))) *
+                  G(90, eval_params.mobilities[p - 2]);)
 
-                    G(90, // KING ATTACKS
+                    G(89, // KING ATTACKS
                       score +=
-                      G(92, count(G(93, mobility) & G(93, opp_king_zone))) *
-                      G(92, eval_params.king_attacks[p - 2]);))
+                      G(91, count(G(92, mobility) & G(92, opp_king_zone))) *
+                      G(91, eval_params.king_attacks[p - 2]);))
             })
 
         G(
             62, // PASSED PAWNS
-            if (G(94,
-                  !(G(95, (0x101010101010101ULL << sq)) & G(95, no_passers))) &&
-                G(94, p == Pawn)) {
+            if (G(93,
+                  !(G(94, (0x101010101010101ULL << sq)) & G(94, no_passers))) &&
+                G(93, p == Pawn)) {
               G(
-                  96, if (north(1ULL << sq) & pos->colour[1]) {
+                  95, if (north(1ULL << sq) & pos->colour[1]) {
                     score += eval_params.passed_blocked_pawns[rank - 1];
                   })
 
-              G(96, score += eval_params.passed_pawns[rank - 1];)
+              G(95, score += eval_params.passed_pawns[rank - 1];)
             })
       }
     }
@@ -1074,32 +1053,32 @@ S(1) i32 eval(Position *const restrict pos) {
 }
 
 typedef struct [[nodiscard]] {
-  G(70, Move moves[max_moves];)
-  G(70, i32 static_eval;)
-  G(70, Move killer;)
-  G(70, i32 num_moves;)
-  G(70, u64 position_hash;)
-  G(70, Move best_move;)
+  G(71, i32 static_eval;)
+  G(71, Move killer;)
+  G(71, Move best_move;)
+  G(71, u64 position_hash;)
+  G(71, i32 num_moves;)
+  G(71, Move moves[max_moves];)
 } SearchStack;
 
 typedef struct [[nodiscard]] __attribute__((packed)) {
-  G(97, i16 score;)
-  G(97, i8 depth;)
-  G(97, Move move;)
-  G(97, u16 partial_hash;)
-  G(97, u8 flag;)
+  G(96, u16 partial_hash;)
+  G(96, i16 score;)
+  G(96, Move move;)
+  G(96, i8 depth;)
+  G(96, u8 flag;)
 } TTEntry;
 _Static_assert(sizeof(TTEntry) == 10);
 
 enum { tt_length = 64 * 1024 * 1024 / sizeof(TTEntry) };
 enum { Upper = 0, Lower = 1, Exact = 2 };
 enum { max_ply = 96 };
-enum { mate = 30720, inf = 31744 };
+enum { mate = 31744, inf = 32256 };
 
-G(98, S(1) i32 move_history[2][6][64][64];)
-G(98, S(0) size_t max_time;)
-G(98, S(0) size_t start_time;)
-G(98, S(1) TTEntry tt[tt_length];)
+G(97, S(1) i32 move_history[2][6][64][64];)
+G(97, S(0) size_t max_time;)
+G(97, S(0) size_t start_time;)
+G(97, S(1) TTEntry tt[tt_length];)
 
 #if defined(__x86_64__) || defined(_M_X64)
 typedef long long __attribute__((__vector_size__(16))) i128;
@@ -1156,14 +1135,14 @@ get_hash(const Position *const pos) {
 #endif
 
 S(1)
-i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
-           H(99, 1, const i32 ply), H(99, 1, i32 depth),
-           H(100, 1, const bool do_null),
+i16 search(H(98, 1, Position *const restrict pos), H(98, 1, i32 alpha),
+           H(98, 1, const i32 ply), H(98, 1, i32 depth),
+           H(99, 1, SearchStack *restrict stack),
 #ifdef FULL
            u64 *nodes,
 #endif
-           H(100, 1, const i32 pos_history_count), H(100, 1, const i32 beta),
-           H(100, 1, SearchStack *restrict stack)) {
+           H(99, 1, const i32 beta), H(99, 1, const i32 pos_history_count),
+           H(99, 1, const bool do_null)) {
   assert(alpha < beta);
   assert(ply >= 0);
 
@@ -1174,32 +1153,28 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
     depth++;
   }
 
-  // EARLY EXITS
-  if (depth > 4 && get_time() - start_time > max_time) {
-    return alpha;
-  }
-
   const u64 tt_hash = get_hash(pos);
 
   // FULL REPETITION DETECTION
   bool in_qsearch = depth <= 0;
-  for (i32 i = G(101, ply) + G(101, pos_history_count);
-       G(102, i > 0) && G(102, do_null); i -= 2) {
+  for (i32 i = G(100, ply) + G(100, pos_history_count);
+       G(101, i > 0) && G(101, do_null); i -= 2) {
     if (tt_hash == stack[i].position_hash) {
       return 0;
     }
   }
 
   // TT PROBING
-  G(103, stack[ply].best_move = (Move){0};)
-  G(103, TTEntry *tt_entry = &tt[tt_hash % tt_length];)
-  G(103, const u16 tt_hash_partial = tt_hash / tt_length;)
+  G(102, stack[ply].best_move = (Move){0};)
+  G(102, const u16 tt_hash_partial = tt_hash / tt_length;)
+  G(102, TTEntry *tt_entry = &tt[tt_hash % tt_length];)
   if (tt_entry->partial_hash == tt_hash_partial) {
     stack[ply].best_move = tt_entry->move;
 
     // TT PRUNING
-    if (G(104, alpha == beta - 1) && G(104, tt_entry->depth >= depth) &&
-        G(104, tt_entry->flag != tt_entry->score <= alpha)) {
+    if (G(103, alpha == beta - 1) &&
+        G(103, tt_entry->flag != tt_entry->score <= alpha) &&
+        G(103, tt_entry->depth >= depth)) {
       return tt_entry->score;
     }
   } else if (depth > 3) {
@@ -1210,104 +1185,107 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
 
   // STATIC EVAL WITH ADJUSTMENT FROM TT
   i32 static_eval = eval(pos);
+  assert(static_eval < mate);
+  assert(static_eval > -mate);
+
   stack[ply].static_eval = static_eval;
   const bool improving = ply > 1 && static_eval > stack[ply - 2].static_eval;
-  if (G(105, tt_entry->partial_hash == tt_hash_partial) &&
-      G(105, tt_entry->flag != static_eval > tt_entry->score)) {
+  if (G(104, tt_entry->flag != static_eval > tt_entry->score) &&
+      G(104, tt_entry->partial_hash == tt_hash_partial)) {
     static_eval = tt_entry->score;
   }
 
   // QUIESCENCE
-  if (G(106, static_eval > alpha) && G(106, in_qsearch)) {
+  if (G(105, static_eval > alpha) && G(105, in_qsearch)) {
     if (static_eval >= beta) {
       return static_eval;
     }
     alpha = static_eval;
   }
 
-  if (G(107, !in_check) && G(107, alpha == beta - 1)) {
-    if (G(108, depth < 8) && G(108, !in_qsearch)) {
+  if (G(106, !in_check) && G(106, alpha == beta - 1)) {
+    if (G(107, depth < 8) && G(107, !in_qsearch)) {
 
-      G(109, {
+      G(108, {
         // REVERSE FUTILITY PRUNING
         if (static_eval - 52 * depth >= beta) {
           return static_eval;
         }
       })
 
-      G(109, // RAZORING
+      G(108, // RAZORING
         in_qsearch = static_eval + 123 * depth <= alpha;)
     }
 
     // NULL MOVE PRUNING
-    if (G(110, depth > 2) && G(110, static_eval >= beta) && G(110, do_null)) {
+    if (G(109, depth > 2) && G(109, static_eval >= beta) && G(109, do_null)) {
       Position npos = *pos;
-      G(111, flip_pos(&npos);)
-      G(111, npos.ep = 0;)
+      G(110, npos.ep = 0;)
+      G(110, flip_pos(&npos);)
       const i32 score = -search(
-          H(99, 2, &npos), H(99, 2, -beta), H(99, 2, ply + 1),
-          H(99, 2, depth - 3 - depth / 4), H(100, 2, false),
+          H(98, 2, &npos), H(98, 2, -beta), H(98, 2, ply + 1),
+          H(98, 2, depth - 3 - depth / 4), H(99, 2, stack),
 #ifdef FULL
           nodes,
 #endif
-          H(100, 2, pos_history_count), H(100, 2, -alpha), H(100, 2, stack));
+          H(99, 2, -alpha), H(99, 2, pos_history_count), H(99, 2, false));
       if (score >= beta) {
         return score;
       }
     }
   }
 
-  G(99, i32 quiets_evaluated = 0;)
-  G(99, i32 moves_evaluated = 0;)
-  G(99, stack[pos_history_count + ply + 2].position_hash = tt_hash;)
-  G(99, stack[ply].num_moves = movegen(
+  G(98, i32 quiets_evaluated = 0;)
+  G(98, i32 moves_evaluated = 0;)
+  G(98, stack[ply].num_moves = movegen(
             H(63, 3, pos), H(63, 3, stack[ply].moves), H(63, 3, in_qsearch));)
-  G(99, i32 best_score = in_qsearch ? static_eval : -inf;)
-  G(99, u8 tt_flag = Upper;)
+  G(98, u8 tt_flag = Upper;)
+  G(98, i32 best_score = in_qsearch ? static_eval : -inf;)
+  G(98, stack[pos_history_count + ply + 2].position_hash = tt_hash;)
 
   for (i32 move_index = 0; move_index < stack[ply].num_moves; move_index++) {
     // MOVE ORDERING
-    G(112, i32 move_score = ~0x1010101LL;)
-    G(112, i32 best_index = 0;)
+    G(111, i32 move_score = ~0x1010101LL;)
+    G(111, i32 best_index = 0;)
     for (i32 order_index = move_index; order_index < stack[ply].num_moves;
          order_index++) {
       assert(
           stack[ply].moves[order_index].takes_piece ==
           piece_on(H(31, 7, pos), H(31, 7, stack[ply].moves[order_index].to)));
       const i32 order_move_score =
-          G(100, // KILLER MOVE
-            move_equal(G(113, &stack[ply].moves[order_index]),
-                       G(113, &stack[ply].killer)) *
+          G(99, // KILLER MOVE
+            move_equal(G(112, &stack[ply].moves[order_index]),
+                       G(112, &stack[ply].killer)) *
                 861) +
-          G(100, // PREVIOUS BEST MOVE FIRST
-            (move_equal(G(114, &stack[ply].best_move),
-                        G(114, &stack[ply].moves[order_index]))
+          G(99, // PREVIOUS BEST MOVE FIRST
+            (move_equal(G(113, &stack[ply].best_move),
+                        G(113, &stack[ply].moves[order_index]))
              << 30)) +
-          G(100, // MOST VALUABLE VICTIM
+          G(99, // MOST VALUABLE VICTIM
             stack[ply].moves[order_index].takes_piece * 737) +
-          G(100, // HISTORY HEURISTIC
+          G(99, // HISTORY HEURISTIC
             move_history[pos->flipped]
                         [stack[ply].moves[order_index].takes_piece]
                         [stack[ply].moves[order_index].from]
                         [stack[ply].moves[order_index].to]);
       if (order_move_score > move_score) {
-        G(115, best_index = order_index;)
-        G(115, move_score = order_move_score;)
+        G(114, best_index = order_index;)
+        G(114, move_score = order_move_score;)
       }
     }
 
-    swapmoves(G(116, &stack[ply].moves[best_index]),
-              G(116, &stack[ply].moves[move_index]));
+    swapmoves(G(115, &stack[ply].moves[move_index]),
+              G(115, &stack[ply].moves[best_index]));
 
     // FORWARD FUTILITY PRUNING / DELTA PRUNING
-    if (G(117, depth < 8) &&
-        G(117,
-          G(118, static_eval + 136 * depth) +
-                  G(118, max_material[stack[ply].moves[move_index].promo]) +
-                  G(118,
+    if (G(116, depth < 8) &&
+        G(116,
+          G(117, static_eval + 136 * depth) +
+                  G(117, max_material[stack[ply].moves[move_index].promo]) +
+                  G(117,
                     max_material[stack[ply].moves[move_index].takes_piece]) <
               alpha) &&
-        G(117, moves_evaluated) && G(117, !in_check)) {
+        G(116, moves_evaluated) && G(116, !in_check)) {
       break;
     }
 
@@ -1324,20 +1302,25 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
     moves_evaluated++;
 
     // LATE MOVE REDUCTION
-    i32 reduction = G(119, depth > 1) && G(119, moves_evaluated > 6)
-                        ? G(120, (alpha == beta - 1)) +
-                              G(120, moves_evaluated / 11) + G(120, !improving)
+    i32 reduction = G(118, depth > 1) && G(118, moves_evaluated > 6)
+                        ? G(119, (alpha == beta - 1)) +
+                              G(119, moves_evaluated / 11) + G(119, !improving)
                         : 0;
 
     i32 score;
     while (true) {
       score = -search(
-          H(99, 3, &npos), H(99, 3, low), H(99, 3, ply + 1),
-          H(99, 3, depth - G(121, 1) - G(121, reduction)), H(100, 3, true),
+          H(98, 3, &npos), H(98, 3, low), H(98, 3, ply + 1),
+          H(98, 3, depth - G(120, 1) - G(120, reduction)), H(99, 3, stack),
 #ifdef FULL
           nodes,
 #endif
-          H(100, 3, pos_history_count), H(100, 3, -alpha), H(100, 3, stack));
+          H(99, 3, -alpha), H(99, 3, pos_history_count), H(99, 3, true));
+
+      // EARLY EXITS
+      if (depth > 4 && get_time() - start_time > max_time) {
+        return best_score;
+      }
 
       if (score > alpha) {
         if (reduction != 0) {
@@ -1357,21 +1340,21 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
       best_score = score;
 
       if (score > alpha) {
-        G(122, tt_flag = Exact;)
-        G(122, stack[ply].best_move = stack[ply].moves[move_index];)
-        G(122, alpha = score;)
+        G(121, tt_flag = Exact;)
+        G(121, stack[ply].best_move = stack[ply].moves[move_index];)
+        G(121, alpha = score;)
         if (score >= beta) {
           assert(stack[ply].best_move.takes_piece ==
                  piece_on(H(31, 8, pos), H(31, 8, stack[ply].best_move.to)));
-          G(123, tt_flag = Lower;)
+          G(122, tt_flag = Lower;)
           G(
-              123, if (stack[ply].best_move.takes_piece == None) {
+              122, if (stack[ply].best_move.takes_piece == None) {
                 stack[ply].killer = stack[ply].best_move;
               })
           G(
-              123, if (!in_qsearch) {
+              122, if (!in_qsearch) {
                 const i32 bonus = depth * depth;
-                G(124, i32 *const this_hist =
+                G(123, i32 *const this_hist =
                            &move_history[pos->flipped]
                                         [stack[ply].best_move.takes_piece]
                                         [stack[ply].best_move.from]
@@ -1379,7 +1362,7 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
 
                   *this_hist += bonus - bonus * *this_hist / 1024;)
                 G(
-                    124, for (i32 prev_index = 0; prev_index < move_index;
+                    123, for (i32 prev_index = 0; prev_index < move_index;
                               prev_index++) {
                       const Move prev = stack[ply].moves[prev_index];
                       i32 *const prev_hist =
@@ -1398,8 +1381,8 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
     }
 
     // LATE MOVE PRUNING
-    if (G(125, !in_check) && G(125, alpha == beta - 1) &&
-        G(125, quiets_evaluated > 1 + depth * depth >> !improving)) {
+    if (G(124, !in_check) && G(124, alpha == beta - 1) &&
+        G(124, quiets_evaluated > 1 + depth * depth >> !improving)) {
       break;
     }
   }
@@ -1418,6 +1401,40 @@ i16 search(H(99, 1, Position *const restrict pos), H(99, 1, i32 alpha),
   return best_score;
 }
 
+S(1) void init() {
+  G(
+      46,
+      // INIT DIAGONAL MASKS
+      for (i32 sq = 0; sq < 64; sq++) {
+        const u64 bb = 1ULL << sq;
+        diag_mask[sq] =
+            G(125, ray(H(15, 4, 0), H(15, 4, ~0x8080808080808080ull),
+                       H(15, 4, bb), H(15, 4, -9))) | // Northeast
+            G(125, ray(H(15, 5, 0), H(15, 5, ~0x101010101010101ull),
+                       H(15, 5, bb), H(15, 5, 9))); // Southwest
+      })
+  G(
+      46, // MERGE MATERIAL VALUES
+      for (i32 i = 0; i < sizeof(mg.material) / sizeof(i16); i++) {
+        eval_params.material[i] = combine_eval_param(H(74, 2, mg.material[i]),
+                                                     H(74, 2, eg.material[i]));
+      })
+
+  G(46,
+    // CLEAR HISTORY
+    __builtin_memset(move_history, 0, sizeof(move_history));)
+  G(
+      46, // MERGE NON-MATERIAL VALUES
+      for (i32 i = 0; i < sizeof(mg) - sizeof(mg.material); i++) {
+        // Technically writes past end of array
+        // But since the structs are packed, it works
+        const i32 offset = sizeof(mg.material);
+        ((i32 *)&eval_params)[offset / sizeof(*mg.material) + i] =
+            combine_eval_param(H(74, 3, ((i8 *)&mg)[offset + i]),
+                               H(74, 3, ((i8 *)&eg)[offset + i]));
+      })
+}
+
 S(1)
 void iteratively_deepen(
 #ifdef FULL
@@ -1433,15 +1450,20 @@ void iteratively_deepen(
   for (i32 depth = 1; depth < max_ply; depth++) {
 #endif
     i32 score =
-        search(H(99, 4, pos), H(99, 4, -inf), H(99, 4, 0), H(99, 4, depth),
-               H(100, 4, false),
+        search(H(98, 4, pos), H(98, 4, -inf), H(98, 4, 0), H(98, 4, depth),
+               H(99, 4, stack),
 #ifdef FULL
                nodes,
 #endif
-               H(100, 4, pos_history_count), H(100, 4, inf), H(100, 4, stack));
-    size_t elapsed = get_time() - start_time;
+               H(99, 4, inf), H(99, 4, pos_history_count), H(99, 4, false));
 
+    size_t elapsed = get_time() - start_time;
 #ifdef FULL
+    // Don't print unreliable scores
+    if (elapsed > max_time) {
+      break;
+    }
+
     printf("info depth %i score cp %i time %i nodes %i", depth, score, elapsed,
            *nodes);
     if (elapsed > 0) {
@@ -1580,14 +1602,13 @@ S(1) void run() {
   G(127, char line[4096];)
   G(127, Position pos;)
   G(127, i32 pos_history_count;)
-  G(127, init();)
-  G(127, __builtin_memset(move_history, 0, sizeof(move_history));)
   G(127, // #ifdef LOWSTACK
          //  SearchStack *stack = malloc(sizeof(SearchStack) * 1024);
          // #else
     SearchStack stack[1024];
     // #endif
   )
+  G(127, init();)
 
 #ifdef FULL
   pos = start_pos;
@@ -1636,44 +1657,44 @@ S(1) void run() {
              nps);
     }
 #endif
-    G(128, if (line[0] == 'i') { puts("readyok"); })
-    else G(128, if (line[0] == 'q') { exit_now(); }) else if (line[0] == 'p') {
+    G(128, if (line[0] == 'q') { exit_now(); })
+    else G(128, if (line[0] == 'p') {
       G(129, pos_history_count = 0;)
-      G(129, pos = start_pos;)
-      while (true) {
-        const bool line_continue = getl(line);
+        G(129, pos = start_pos;)
+        while (true) {
+          const bool line_continue = getl(line);
 
 #if FULL
-        if (!strcmp(line, "fen")) {
-          getl(line);
-          get_fen(&pos, line);
-        }
+          if (!strcmp(line, "fen")) {
+            getl(line);
+            get_fen(&pos, line);
+          }
 #endif
 
-        const i32 num_moves =
+          const i32 num_moves =
             movegen(H(63, 4, &pos), H(63, 4, stack[0].moves), H(63, 4, false));
-        for (i32 i = 0; i < num_moves; i++) {
-          char move_name[8];
-          move_str(H(28, 4, move_name), H(28, 4, &stack[0].moves[i]),
-                   H(28, 4, pos.flipped));
-          assert(move_string_equal(line, move_name) ==
-                 !strcmp(line, move_name));
-          if (move_string_equal(G(130, line), G(130, move_name))) {
-            stack[pos_history_count].position_hash = get_hash(&pos);
-            pos_history_count++;
-            if (stack[0].moves[i].takes_piece != None) {
-              pos_history_count = 0;
+          for (i32 i = 0; i < num_moves; i++) {
+            char move_name[8];
+            move_str(H(28, 4, move_name), H(28, 4, &stack[0].moves[i]),
+              H(28, 4, pos.flipped));
+            assert(move_string_equal(line, move_name) ==
+              !strcmp(line, move_name));
+            if (move_string_equal(G(130, line), G(130, move_name))) {
+              stack[pos_history_count].position_hash = get_hash(&pos);
+              pos_history_count++;
+              if (stack[0].moves[i].takes_piece != None) {
+                pos_history_count = 0;
+              }
+              makemove(H(46, 4, &pos), H(46, 4, &stack[0].moves[i]));
+              break;
             }
-            makemove(H(46, 4, &pos), H(46, 4, &stack[0].moves[i]));
+          }
+          if (!line_continue) {
             break;
           }
         }
-        if (!line_continue) {
-          break;
-        }
-      }
-    }
-    else if (line[0] == 'g') {
+    })
+    else G(128, if (line[0] == 'g') {
 #ifdef FULL
       while (true) {
         getl(line);
@@ -1681,26 +1702,29 @@ S(1) void run() {
           getl(line);
           max_time = atoi(line) / 2;
           break;
-        } else if (pos.flipped && !strcmp(line, "btime")) {
+        }
+        else if (pos.flipped && !strcmp(line, "btime")) {
           getl(line);
           max_time = atoi(line) / 2;
           break;
-        } else if (!strcmp(line, "movetime")) {
+        }
+        else if (!strcmp(line, "movetime")) {
           max_time = 20000; // Assume Lichess bot
           break;
         }
       }
       iteratively_deepen(max_ply, &nodes, H(126, 4, &pos), H(126, 4, stack),
-                         H(126, 4, pos_history_count));
+        H(126, 4, pos_history_count));
 #else
       for (i32 i = 0; i < (pos.flipped ? 4 : 2); i++) {
         getl(line);
         max_time = atoi(line) / 2;
       }
       iteratively_deepen(H(126, 5, &pos), H(126, 5, stack),
-                         H(126, 5, pos_history_count));
+        H(126, 5, pos_history_count));
 #endif
-    }
+    })
+    else G(128, if (line[0] == 'i') { puts("readyok"); })
   }
 }